+++ lib/appsignal/hooks/mongo_ruby_driver.rb
@@ -16,6 +16,5 @@ module Appsignal
         )
       end
     end
-
   end
 end