+++ spec/lib/appsignal_spec.rb
@@ -101,6 +101,11 @@ describe Appsignal do
           Appsignal::Extension.should_receive(:install_gc_event_hooks)
           Appsignal.start
         end
+
+        it "should add the gc probe to minutely" do
+          Appsignal::Minutely.should_receive(:add_gc_probe)
+          Appsignal.start
+        end
       end
 
       context "when allocation tracking and gc instrumentation have been disabled" do
@@ -112,6 +117,33 @@ describe Appsignal do
         it "should not install event hooks" do
           Appsignal::Extension.should_not_receive(:install_allocation_event_hook)
           Appsignal::Extension.should_not_receive(:install_gc_event_hooks)
+          Appsignal.start
+        end
+
+       it "should not add the gc probe to minutely" do
+          Appsignal::Minutely.should_not_receive(:add_gc_probe)
+          Appsignal.start
+       end
+      end
+
+      context "when minutely metrics has been enabled" do
+        before do
+          Appsignal.config.config_hash[:enable_minutely_probes] = true
+        end
+
+        it "should start minutely" do
+          Appsignal::Minutely.should_receive(:start)
+          Appsignal.start
+        end
+      end
+
+      context "when minutely metrics has been disabled" do
+        before do
+          Appsignal.config.config_hash[:enable_minutely_probes] = false
+        end
+
+        it "should not start minutely" do
+          Appsignal::Minutely.should_not_receive(:start)
           Appsignal.start
         end
       end
@@ -266,6 +298,19 @@ describe Appsignal do
         }.should_not raise_error
       end
     end
+
+    describe ".instrument" do
+      it "should not instrument, but still call the block" do
+        stub = double
+        stub.should_receive(:method_call).and_return('return value')
+
+        lambda {
+          Appsignal.instrument 'name' do
+            stub.method_call
+          end.should eq 'return value'
+        }.should_not raise_error
+      end
+    end
   end
 
   context "with config and started" do
@@ -393,14 +438,14 @@ describe Appsignal do
 
     describe "custom stats" do
       describe ".set_gauge" do
-        it "should call set_gauge on the extension with a float" do
+        it "should call set_gauge on the extension with a string key and float" do
           Appsignal::Extension.should_receive(:set_gauge).with('key', 0.1)
           Appsignal.set_gauge('key', 0.1)
         end
 
-        it "should call set_gauge on the extension with an int" do
+        it "should call set_gauge on the extension with a symbol key and int" do
           Appsignal::Extension.should_receive(:set_gauge).with('key', 1.0)
-          Appsignal.set_gauge('key', 1)
+          Appsignal.set_gauge(:key, 1)
         end
 
         it "should not raise an exception when out of range" do
@@ -413,14 +458,14 @@ describe Appsignal do
       end
 
       describe ".set_host_gauge" do
-        it "should call set_host_gauge on the extension with a float" do
+        it "should call set_host_gauge on the extension with a string key and float" do
           Appsignal::Extension.should_receive(:set_host_gauge).with('key', 0.1)
           Appsignal.set_host_gauge('key', 0.1)
         end
 
-        it "should call set_host_gauge on the extension with an int" do
+        it "should call set_host_gauge on the extension with a symbol key and int" do
           Appsignal::Extension.should_receive(:set_host_gauge).with('key', 1.0)
-          Appsignal.set_host_gauge('key', 1)
+          Appsignal.set_host_gauge(:key, 1)
         end
 
         it "should not raise an exception when out of range" do
@@ -433,14 +478,14 @@ describe Appsignal do
       end
 
       describe ".set_process_gauge" do
-        it "should call set_process_gauge on the extension with a float" do
+        it "should call set_process_gauge on the extension with a string key and float" do
           Appsignal::Extension.should_receive(:set_process_gauge).with('key', 0.1)
           Appsignal.set_process_gauge('key', 0.1)
         end
 
-        it "should call set_process_gauge on the extension with an int" do
+        it "should call set_process_gauge on the extension with a symbol key and int" do
           Appsignal::Extension.should_receive(:set_process_gauge).with('key', 1.0)
-          Appsignal.set_process_gauge('key', 1)
+          Appsignal.set_process_gauge(:key, 1)
         end
 
         it "should not raise an exception when out of range" do
@@ -453,9 +498,14 @@ describe Appsignal do
       end
 
       describe ".increment_counter" do
-        it "should call increment_counter on the extension" do
+        it "should call increment_counter on the extension with a string key" do
           Appsignal::Extension.should_receive(:increment_counter).with('key', 1)
           Appsignal.increment_counter('key')
+        end
+
+        it "should call increment_counter on the extension with a symbol key" do
+          Appsignal::Extension.should_receive(:increment_counter).with('key', 1)
+          Appsignal.increment_counter(:key)
         end
 
         it "should call increment_counter on the extension with a count" do
@@ -473,14 +523,14 @@ describe Appsignal do
       end
 
       describe ".add_distribution_value" do
-        it "should call add_distribution_value on the extension with a float" do
+        it "should call add_distribution_value on the extension with a string key and float" do
           Appsignal::Extension.should_receive(:add_distribution_value).with('key', 0.1)
           Appsignal.add_distribution_value('key', 0.1)
         end
 
-        it "should call add_distribution_value on the extension with an int" do
+        it "should call add_distribution_value on the extension with a symbol key and int" do
           Appsignal::Extension.should_receive(:add_distribution_value).with('key', 1.0)
-          Appsignal.add_distribution_value('key', 1)
+          Appsignal.add_distribution_value(:key, 1)
         end
 
         it "should not raise an exception when out of range" do
@@ -673,6 +723,25 @@ describe Appsignal do
         transaction.should_not_receive(:set_error)
 
         Appsignal.set_error(nil)
+      end
+    end
+
+    describe ".instrument" do
+      it "should instrument through the transaction" do
+        stub = double
+        stub.should_receive(:method_call).and_return('return value')
+
+        transaction.should_receive(:start_event)
+        transaction.should_receive(:finish_event).with(
+          'name',
+          'title',
+          'body',
+          0
+        )
+
+        Appsignal.instrument 'name', 'title', 'body' do
+          stub.method_call
+        end.should eq 'return value'
       end
     end